src: ProjectionThread: Keep track on matched points
This commit adds a matrix to determine which points has been
matched/fused, so these don't triangulate/add new points to the map.

We have to keep this info in a different matrix, instead of setting
that pixel disparity as invalid because we may want to compare more
points in the map with that pixel.

Signed-off-by: Ariel D'Alessandro <ariel@vanguardiasur.com.ar>