src: ProjectionThread: Don't discard confirmed points
Those points that has already been confirmed, i.e. matched/merged at
least once, are not going to be discarded now.

This commit also adds the number of merged points data to the log file.
This is the number of points, that are part of the final point cloud,
which have been matched/merged at least one time.

Signed-off-by: Ariel D'Alessandro <ariel@vanguardiasur.com.ar>