Remove single_depth_map_region_size parameter
This parameter is not needed in the dense node. So, let's add it as an
argument in the dmap tools, where it is really required.

While here, remove the single_cloud_path parameter/functionality
provided by the dense node, which was implemented as a special mode to
publish a point cloud message and exit. This should be implemented as a
single separated tool, useful to visualize point clouds in rviz (can
this be achieved in a different way?).

Signed-off-by: Ariel D'Alessandro <ariel@vanguardiasur.com.ar>